backend/requirements.txt
Django==4.1
djoser==2.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==23.1
PyJWT==2.1.0
drf-extra-fields==3.4.0
flake8==6.0.0